+++ Src/FluentAssertions/Equivalency/Steps/StringEqualityEquivalencyStep.cs
@@ -79,7 +79,7 @@ public class StringEqualityEquivalencyStep : IEquivalencyStep
         if (onlyOneNull)
         {
             assertionChain.FailWith(
-                $"Expected {currentNode.Subject.Description} to be {{0}}{{reason}}, but found {{1}}.", expected, subject);
+                "Expected {0} to be {1}{reason}, but found {2}.", currentNode.Subject.Description.AsNonFormattable(), expected, subject);
 
             return false;
         }
@@ -95,7 +95,7 @@ public class StringEqualityEquivalencyStep : IEquivalencyStep
         }
 
         assertionChain.FailWith(
-            $"Expected {currentNode} to be {{0}}, but found {{1}}.",
+            "Expected {0} to be {1}, but found {2}.", currentNode.AsNonFormattable(),
             comparands.RuntimeType, comparands.Subject.GetType());
 
         return assertionChain.Succeeded;